requirements.txt
openml==0.12.2
pandas==1.4.3
pyyaml==6.0
matplotlib==3.5.2
numpy==1.23.0
scipy==1.8.1
giotto-tda==0.5.1
networkx==2.8.4
path==16.4.0
tqdm==4.64.0
mlflow==1.27.0
scikit-learn==1.1.1
joblib==1.1.0
python-dotenv==0.20.0
pytest==7.1.2
